Run an alarm for the created tab if the tab is inactive
There is a case that a tab is opened with inactive. (e.g. when clicking a link with the Cmd key.)
The alarm will never be run unless the tab is activated. In other words, the tab will never be closed automatically.
Therefore, made it run the alarm if the inactive tab is created.