Change calculation method for the time left of the tab
If the time left is calculated from the current time and the last
evacuated time (lastEvacuatedAt), it will not be possible to calculate
the correct time left when a new alarm is added.
This change re-calculates the scheduled time
from the current time and pre-calculated time left when recovering
alarms.